refactor: move the request backstop plumbing into its own module
The window helper, the context symbols, and the race that backs the internal
request timeout have grown enough to live on their own. Move them into
`request-backstop.ts` as `raceWithBackstop`, `remainingNavigationWindowMillis`,
and the `navigationDeadline` / `extendBackstop` / `backstopExpired` symbols;
`BasicCrawler` keeps only the crawler-specific sizing (which phases the backstop
must outlast) and delegates the rest. No behaviour change.
